better type

+++ RRGame/Classes/Support/GKTurnBasedMatch/GKTurnBasedMatch+RRHeredox.h
@@ -36,8 +36,8 @@ typedef struct RRMatchData {
 
 @interface GKTurnBasedMatch (RRMatchData)
 
-@property(nonatomic, assign)    NSUInteger     gameSeed;
-@property(nonatomic, assign)    RRPlayerColor  firstParticipantColor;
+@property(nonatomic, assign)    unsigned int    gameSeed;
+@property(nonatomic, assign)    RRPlayerColor   firstParticipantColor;
 
 - (void)addTileMove:(RRTileMove)tileMove;
 - (GKTurnBasedParticipant *)participantForColor:(RRPlayerColor)playerColor;
